completed labelbox.py
